+++ Neillans.Adapters.Secrets.Core/SecretsProviderFactory.cs
@@ -20,6 +20,8 @@ public class SecretsProviderFactory : ISecretsProviderFactory
         {
             SecretsProviderType.AzureKeyVault => _serviceProvider.GetRequiredService<ISecretsProvider>(),
             SecretsProviderType.Infisical => _serviceProvider.GetRequiredService<ISecretsProvider>(),
+            SecretsProviderType.BitWarden => _serviceProvider.GetRequiredService<ISecretsProvider>(),
+            SecretsProviderType.InMemory => _serviceProvider.GetRequiredService<ISecretsProvider>(),
             _ => throw new NotSupportedException($"Provider type '{providerType}' is not supported.")
         };
     }